+++ JuliaBUGS/src/model/Model.jl
@@ -2,8 +2,10 @@ module Model
 
 using Accessors
 using AbstractPPL
+using ADTypes
 using BangBang
 using Bijectors
+import DifferentiationInterface as DI
 using Distributions
 using Graphs
 using LinearAlgebra
@@ -25,6 +27,9 @@ export set_evaluation_mode, set_observed_values!
 
 # Evaluation mode types
 export UseGraph, UseGeneratedLogDensityFunction, UseAutoMarginalization
+
+# Gradient wrapper
+export BUGSModelWithGradient
 
 # Internal evaluation functions (exported for testing, not re-exported to users)
 export evaluate_with_rng!!, evaluate_with_env!!, evaluate_with_values!!